requirements.txt
--index-url https://download.pytorch.org/whl/cu118
--extra-index-url https://pypi.org/simple
pillow
torch
torchvision